+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [0.0.7] - 2022-04-01
 
 ### Added
@@ -12,4 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Option to exclude regular expressions from being counted as words but keep them in the content
 - Added a button to toggle cards display on Writing Plan Outline Tree.
 - Added Writing Plan Support for Cabinet Preview;
-.
+- Generate footnote from markdown